Change offset type to Long in RdfKafkaOffsetRepository
Offset stored in Blazegraph was parsed as Integer
which caused the parsing to fail when actual offset
went over Integer.MAX_VALUE. This patch switches the type
to Long.

+++ tools/src/main/java/org/wikidata/query/rdf/tool/change/RdfKafkaOffsetsRepository.java
@@ -62,7 +62,7 @@ public class RdfKafkaOffsetsRepository implements KafkaOffsetsRepository {
                             String[] parts = e.getKey().split(":", 2);
                             return new TopicPartition(parts[0], Integer.parseInt(parts[1]));
                         },
-                        e -> new OffsetAndTimestamp(Integer.parseInt(e.getValue()), firstStartTime.toEpochMilli())
+                        e -> new OffsetAndTimestamp(Long.parseLong(e.getValue()), firstStartTime.toEpochMilli())
                 ));
     }
